job优化

+++ windrises-core/src/main/java/com/windrises/core/entity/po/QuartzLog.java
@@ -56,7 +56,7 @@ public class QuartzLog extends BasePo {
     /**
      * 耗时（毫秒）
      */
-    private Integer elapsedTime;
+    private Long elapsedTime;
 
     /**
      * 创建日期
